refactor: move EmissionLine + LINE_LAMBDAS to kl_pipe/lines.py
EmissionLine and the rest-wavelength registry are spectral-feature
concepts, not source-container concepts. Conceptual home is alongside
spectral machinery; cannot land in spectral.py without colliding with
the legacy `EmissionLine = _LegacyEmissionLine` back-compat alias.

New dedicated module:
- kl_pipe/lines.py owns LINE_LAMBDAS and the new EmissionLine dataclass
  (incl. pytree registration). Independent of SpectralModel; survives
  the eventual SpectralModel deletion intact.

source.py drops the EmissionLine/LINE_LAMBDAS definitions and imports
them from kl_pipe.lines for use in SourceModel.

Import-path updates:
- tests/test_source.py: EmissionLine + LINE_LAMBDAS now from kl_pipe.lines.
- SOURCE_MODEL_EXAMPLES.py: same.

53 SourceModel/coordinates/obs-rotation tests green; existing imports
of `EmissionLine` from `kl_pipe.spectral` (legacy alias path) unchanged.
