+++ kl_pipe/source.py
@@ -14,6 +14,9 @@ parameter namespace (``vel.vcirc``, ``F087.flux``, ``Halpha.flux``, etc.).
 The underlying VelocityModel / IntensityModel classes operate on flat
 theta arrays matching their own ``PARAMETER_NAMES`` and stay unaware of
 the namespace.
+
+``EmissionLine`` and the ``LINE_LAMBDAS`` rest-wavelength registry live
+in ``kl_pipe.lines`` (import from there directly).
 """
 
 from __future__ import annotations
@@ -23,102 +26,10 @@ from typing import TYPE_CHECKING, Dict, Optional
 
 import jax
 
+from kl_pipe.lines import LINE_LAMBDAS, EmissionLine
+
 if TYPE_CHECKING:
     from kl_pipe.model import IntensityModel, VelocityModel
-
-
-# ===========================================================================
-# LINE_LAMBDAS registry: vacuum rest wavelengths (nm)
-# ===========================================================================
-
-LINE_LAMBDAS: Dict[str, float] = {
-    # singlets — canonical name suffices
-    'Lyalpha': 121.567,
-    'CIV': 154.95,
-    'CIII': 190.9,
-    'MgII': 279.8,
-    'Hbeta': 486.13,
-    'Hgamma': 434.05,
-    'Halpha': 656.28,
-    # doublets / multiplets — wavelength integer suffix required
-    'OII3726': 372.60,  # [O II] doublet, weaker
-    'OII3728': 372.88,  # [O II] doublet, stronger
-    'OII3727': 372.74,  # [O II] blended (low-R convention)
-    'OIII4959': 495.89,  # [O III] doublet, weaker
-    'OIII5007': 500.68,  # [O III] doublet, stronger
-    'NII6548': 654.81,  # [N II] doublet, weaker
-    'NII6584': 658.35,  # [N II] doublet, stronger
-    'SII6717': 671.65,  # [S II] doublet, weaker
-    'SII6731': 673.08,  # [S II] doublet, stronger
-}
-
-
-# ===========================================================================
-# EmissionLine
-# ===========================================================================
-
-
-@dataclass
-class EmissionLine:
-    """One emission line in a SourceModel.
-
-    Parameters
-    ----------
-    intensity : IntensityModel, optional
-        Spatial profile of the ionized-gas emission at this line wavelength.
-        Mutually exclusive with ``intensity_key``; exactly one must be set.
-    intensity_key : str, optional
-        Reference to another emission line's ``intensity`` (e.g.
-        ``intensity_key='Halpha'`` to share Halpha's spatial profile).
-        ``<line>.flux`` is still per-line.
-    continuum : IntensityModel, optional
-        Optional stellar continuum at this line's wavelength. Adds a
-        broadband-like component under the line in cube assembly.
-        Mutually exclusive with ``continuum_key``.
-    continuum_key : str, optional
-        Reference to another emission line's ``continuum``. Same sharing
-        semantics as ``intensity_key`` but for the continuum component.
-        ``<line>.cont.flux`` is still per-line.
-    dispersion_key : str, optional
-        Reference to another emission line's intrinsic kinematic
-        velocity dispersion. When set, this line's dispersion is read
-        from the referenced line's ``<line>.dispersion`` prior rather
-        than from its own. Validated by ``SourceModel.__post_init__``.
-    lambda_rest : float, optional
-        Rest-frame line wavelength in nm (vacuum). If None at
-        construction, SourceModel will auto-resolve from ``LINE_LAMBDAS``
-        using this line's dict key.
-    """
-
-    intensity: Optional['IntensityModel'] = None
-    intensity_key: Optional[str] = None
-    continuum: Optional['IntensityModel'] = None
-    continuum_key: Optional[str] = None
-    dispersion_key: Optional[str] = None
-    lambda_rest: Optional[float] = None
-
-    def __post_init__(self):
-        # exactly one of intensity / intensity_key must be set
-        has_intensity = self.intensity is not None
-        has_key = self.intensity_key is not None
-        if has_intensity == has_key:
-            raise ValueError(
-                "EmissionLine: exactly one of 'intensity' or 'intensity_key' "
-                "must be set"
-            )
-        # at most one of continuum / continuum_key
-        if self.continuum is not None and self.continuum_key is not None:
-            raise ValueError(
-                "EmissionLine: at most one of 'continuum' or 'continuum_key' "
-                "may be set"
-            )
-        # dispersion_key cross-reference is validated by SourceModel,
-        # since it needs the full emission_lines dict to check.
-
-
-# ===========================================================================
-# SourceModel
-# ===========================================================================
 
 
 @dataclass
@@ -236,21 +147,12 @@ class SourceModel:
 # JAX pytree registration
 # ===========================================================================
 #
-# SourceModel and EmissionLine are treated as opaque Python objects from
-# JAX's perspective: no traceable children, the instance itself is the aux.
-# Users pass them via ``partial(jit_fn, source=source)`` (the same pattern
-# used today for KLModel), so JAX never needs to flatten them at the leaf
-# level. Registering them as pytrees with this trivial split lets
+# SourceModel is treated as an opaque Python object from JAX's perspective:
+# no traceable children, the instance itself is the aux. Users pass it via
+# ``partial(jit_fn, source=source)``, so JAX never needs to flatten it at
+# the leaf level. Registering it as a pytree with this trivial split lets
 # ``jax.tree_util.tree_flatten`` succeed and gives users a place to extend
 # later if SourceModel ever needs to carry traceable data.
-
-
-def _emission_line_flatten(line):
-    return (), line  # empty children, instance as aux
-
-
-def _emission_line_unflatten(aux, children):
-    return aux
 
 
 def _source_model_flatten(source):
@@ -261,9 +163,6 @@ def _source_model_unflatten(aux, children):
     return aux
 
 
-jax.tree_util.register_pytree_node(
-    EmissionLine, _emission_line_flatten, _emission_line_unflatten
-)
 jax.tree_util.register_pytree_node(
     SourceModel, _source_model_flatten, _source_model_unflatten
 )
